MED: Set the pointers to null after freeing them. MOD: the plug-in passes -1 to AnalyseSignatureMOD's first argument. Set the first argument to long instead of size_t. S3M: we specify the max length of the string in strlcpy; we don't need to set the last value to a null char.
git-svn-id: svn://svn.code.sf.net/p/playerpro/code/branches/PlayerPRO6@480 f4605116-7794-4d0b-9e48-65ca9bfeedfd

+++ MADDriver.source/MADDriver.Source/Import-Export/Mod.c
@@ -88,7 +88,7 @@ Cmd* GetMADCommand( register short PosX, register short	TrackIdX, register PatDa
 }
 #endif
 
-static void AnalyseSignatureMOD( size_t EOFo, OSType temp, short *maxInstru, SInt32 *PatternSize, short *tracksNo, MODDef* aMOD)
+static void AnalyseSignatureMOD( long EOFo, OSType temp, short *maxInstru, SInt32 *PatternSize, short *tracksNo, MODDef* aMOD)
 {
 	SInt32 		test, i;
 	Boolean		result;
